labelpts_v0.2/requirements.txt
opencv_python==4.5.1.48
PyYAML==6.0.1
